+++ contract/contracts/interfaces/IPoSPool.sol
@@ -1,22 +1,36 @@
 //SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
 
-import "./VotePowerQueue.sol";
+import {ParamsControl} from "@confluxfans/contracts/InternalContracts/ParamsControl.sol";
+import {VotePowerQueue} from "../utils/VotePowerQueue.sol";
+import {IVotingEscrow} from "./IVotingEscrow.sol";
 
 interface IPoSPool {
   struct PoolSummary {
-    uint64 available;
+    uint256 available;
     uint256 interest;
     uint256 totalInterest; // total interest of all pools
   }
 
   struct UserSummary {
-    uint64 votes;  // Total votes in PoS system, including locking, locked, unlocking, unlocked
-    uint64 available; // locking + locked
-    uint64 locked;
-    uint64 unlocked;
+    uint256 votes;  // Total votes in PoS system, including locking, locked, unlocking, unlocked
+    uint256 available; // locking + locked
+    uint256 locked;
+    uint256 unlocked;
     uint256 claimedInterest;
     uint256 currentInterest;
+  }
+
+  struct PoolShot {
+    uint256 available;
+    uint256 balance;
+    uint256 blockNumber;
+  } 
+
+  struct UserShot {
+    uint256 available;
+    uint256 accRewardPerCfx;
+    uint256 blockNumber;
   }
 
   // admin functions
@@ -47,4 +61,10 @@ interface IPoSPool {
   function userOutQueue(address account) external view returns (VotePowerQueue.QueueNode[] memory);
   function userInQueue(address account, uint64 offset, uint64 limit) external view returns (VotePowerQueue.QueueNode[] memory);
   function userOutQueue(address account, uint64 offset, uint64 limit) external view returns (VotePowerQueue.QueueNode[] memory);
+
+  function lockForVotePower(uint256 amount, uint256 unlockBlockNumber) external;
+  function castVote(uint64 vote_round, ParamsControl.Vote[] calldata vote_data) external;
+  function userLockInfo(address user) external view returns (IVotingEscrow.LockInfo memory);
+  function votingEscrow() external view returns (address);
+  function userVotePower(address user) external view returns (uint256);
 }