+++ contract/contracts/utils/Proxy1967.sol
@@ -3,7 +3,8 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-contract PoSPoolProxy1967 is ERC1967Proxy, Ownable  {
+// The implementation contract must first inherit from Ownable too
+contract Proxy1967 is ERC1967Proxy, Ownable  {
   // initialize() - "0x8129fc1c"
   constructor(address logic, bytes memory data) ERC1967Proxy(logic, data) {}