+++ contract/scripts/core/conflux.js
@@ -1,7 +1,7 @@
 /* eslint-disable prettier/prettier */
 require('dotenv').config();
 const { Conflux, Drip } = require("js-conflux-sdk");
-const { loadPrivateKey } = require('../utils');
+const { loadPrivateKey } = require('../../utils');
 
 const conflux = new Conflux({
   url: process.env.CFX_RPC_URL,
@@ -10,11 +10,10 @@ const conflux = new Conflux({
 });
 
 let account = conflux.wallet.addPrivateKey(loadPrivateKey());
-console.log('Account address: ', account.address);
+// console.log('Account address: ', account.address);
 
 function logReceipt (receipt, msg) {
-    console.log(`${msg} status: ${receipt.outcomeStatus === 0 ? 'success' : 'failed'}`);
-    console.log('TX hash: ', receipt.transactionHash);
+    console.log(`${msg} status: ${receipt.outcomeStatus === 0 ? 'success' : 'failed'}`, 'Tx hash', receipt.transactionHash);
 }
 
 module.exports = {