+++ contract/CHANGELOG.md
@@ -1,5 +1,18 @@
-CHANGELOG
-===
+# CHANGELOG
+
+## v1.5.0
+
+Change the `TOTAL_TOPIC` of VotingEscrow.sol and eSpacePoolBridge.sol from 3 to 4.
+
+## v1.4.0
+
+Support eSpace pool to lock CFX to obtain voting rights for Conflux Governance.
+
+## v1.3.0
+
+This version introduces the function of locking CFX to obtain voting rights for PoS Pools. After obtaining voting rights, you can participate in Conflux on-chain parameters voting and community governance voting. The means user can obtain PoS staking rewards while participating governance voting.
+
+Check this doc for more details: [v1.3](./docs/v1.3.md)
 
 ## v1.2.0
 
@@ -14,7 +27,7 @@ Note: There is a bug in conflux-rust version before `v2.2.0`, Which may lead som
 
 To upgrade the contract follow steps should be taken:
 
-1. Deploy new pool implemention contract (the new deployed contract add will be print in command): `node bin/pool.js deploy Pool`
+1. Deploy new pool implementation contract (the new deployed contract add will be print in command): `node bin/pool.js deploy Pool`
 2. Upgrade PoS pool contract to new impl: `node bin/pool.js upgradePoolContract cfx:acfejystd9h506nw9639wegtbb48g2zc9axxxxxxxx`
 3. Set the lock period: `node bin/pool.js Pool setLockPeriod 2250000`
 4. Set the unlock period: `node bin/pool.js Pool setUnlockPeriod 176400`